Corrections to pitch binning algorithm, transition from double to float

+++ Arduino/pitch_utils.h
@@ -7,11 +7,11 @@
 #include "tuned_note.h"
 
 
-std::array<double, 88> get_pitch_freqs(double reference);
+std::array<float, 88> get_pitch_freqs(float reference);
 
 std::array<note_name *, 88> get_pitch_names();
 
-tuned_note freq_to_note(double freq, std::array<double, 88> pitch_freqs);
+tuned_note freq_to_note(float freq, std::array<float, 88> pitch_freqs);
 
 
 #endif //EECS_352_PROJECT_PITCH_UTILS_H
